Use `VisitorMeta` to power `TypeInfo` so that it's not a huge isinstance chain.
